fix(T2): Разбиение на ЕТ - Стуртуры данных, методы
