+++ zheleznyakov.mark/T2/main.cpp
@@ -1,6 +1,9 @@
+#include <iostream>
 #include <limits>
 #include <algorithm>
-#include "structures.hpp"
+#include <vector>
+#include <iterator>
+#include "data-struct.hpp"
 
 int main()
 {
