+++ zheleznyakov.mark/T2/separators.cpp
@@ -1,0 +1,35 @@
+#include "separators.hpp"
+
+namespace zheleznyakov
+{
+  std::istream &operator>>(std::istream &in, SeparatorIO &&dest)
+  {
+    std::istream::sentry sentry(in);
+    if (!sentry)
+    {
+      return in;
+    }
+    char c = '0';
+    in >> c;
+    if (in && (c != dest.exp))
+    {
+      in.setstate(std::ios::failbit);
+    }
+    return in;
+  }
+
+  std::istream &operator>>(std::istream &in, LabelIO &&dest)
+  {
+    std::istream::sentry sentry(in);
+    if (!sentry)
+    {
+      return in;
+    }
+    std::string data = "";
+    if ((in >> data) && (data != dest.exp))
+    {
+      in.setstate(std::ios::failbit);
+    }
+    return in;
+  }
+}
