+++ zheleznyakov.mark/T2/separators.hpp
@@ -1,0 +1,17 @@
+#include <iomanip>
+
+namespace zheleznyakov
+{
+  struct SeparatorIO
+  {
+    char exp;
+  };
+
+  struct LabelIO
+  {
+    std::string exp;
+  };
+
+  std::istream &operator>>(std::istream &in, SeparatorIO &&dest);
+  std::istream &operator>>(std::istream &in, LabelIO &&dest);
+}
